refractored code
