+++ node-auth-passport/controllers/index.js
@@ -1,4 +1,3 @@
-
 exports.landing = function(req, res) {
   res.render('index', { title: 'Members' });
 }
