+++ node-auth-passport/controllers/passport_custom.js
@@ -1,0 +1,5 @@
+var passport = require('passport');
+
+exports.authenticate = function(path, msg){
+  return passport.authenticate('local', {failureRedirect: path, failureFlash: msg});
+}
