+++ node-auth-passport/models/user.js
@@ -24,25 +24,25 @@ var UserSchema = mongoose.Schema({
   }
 });
 
-var User = module.exports = mongoose.model('User', UserSchema);
+var User =  mongoose.model('User', UserSchema);
 
-module.exports.getUserById = function(id, callback){
+User.getUserById = function(id, callback){
   User.findById(id, callback);
 }
 
-module.exports.getUserByUsername = function(username, callback){
+User.getUserByUsername = function(username, callback){
   var query = {username: username};
   User.findOne(query, callback);
 }
 
-module.exports.comparePassword = function(candidatePassword, hash, callback){
+User.comparePassword = function(candidatePassword, hash, callback){
   bcrypt.compare(candidatePassword, hash, function(err, isMatch) {
     // res === true 
     callback(null, isMatch);
 });
 }
 
-module.exports.createUser = function(newUser, callback){
+User.createUser = function(newUser, callback){
   bcrypt.genSalt(10, function(err, salt) {
     bcrypt.hash(newUser.password, salt, function(err, hash) {
       newUser.password = hash;
@@ -50,3 +50,5 @@ module.exports.createUser = function(newUser, callback){
     });
   });
 }
+
+module.exports = User;
