--- node-auth-passport/routes/index.js
@@ -1,16 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-/* GET home page. */
-router.get('/', ensureAuthenticated, function(req, res, next) {
-  res.render('index', { title: 'Members' });
-});
-
-function ensureAuthenticated(req, res, next) {
-  if (req.isAuthenticated()){
-    return next();
-  }
-  res.redirect('/users/login');
-}
-
-module.exports = router;
